make txn do not wait in incrservice while the logtail is paused to 1.2 (#19417)
Fix hung. Consider the following timing:

logtail paused
cancel all active pipelines
reset logtail state to normal
When a pipeline in 2 is executing and a new transaction is opened in another goroutine to do something (e.g., update the cache of a auto-increment), the new transaction is blocked until the logtail state returns to normal. This creates a circular wait.

Approved by: @iamlinjunhong, @sukki37